use float for speed

+++ include/omega/audio.h
@@ -38,7 +38,7 @@ typedef struct OMG_Audio {
     bool (*mus_set_volume)(struct OMG_Audio* this, OMG_Music* mus, float volume);
     double (*mus_get_pos)(struct OMG_Audio* this, OMG_Music* mus);
     bool (*mus_set_pos)(struct OMG_Audio* this, OMG_Music* mus, double pos);
-    bool (*mus_set_speed)(struct OMG_Audio* this, OMG_Music* mus, double speed);
+    bool (*mus_set_speed)(struct OMG_Audio* this, OMG_Music* mus, float speed);
     OMG_Sound* (*snd_from_fp)(struct OMG_Audio* this, OMG_Sound* snd, const OMG_String* path);
     bool (*snd_destroy)(struct OMG_Audio* this, OMG_Sound* snd);
     bool (*snd_play)(struct OMG_Audio* this, OMG_Sound* snd, int loops, double fade_in);
@@ -78,5 +78,5 @@ OMG_API bool omg_audio_snd_stop(OMG_Audio* this, OMG_Sound* snd);
 OMG_API OMG_Music* omg_audio_mus_from_fp(OMG_Audio* this, OMG_Music* mus, const OMG_String* path);
 OMG_API double omg_audio_mus_get_pos(OMG_Audio* this, OMG_Music* mus);
 OMG_API bool omg_audio_mus_set_pos(OMG_Audio* this, OMG_Music* mus, double pos);
-OMG_API bool omg_audio_mus_set_speed(OMG_Audio* this, OMG_Music* mus, double speed);
+OMG_API bool omg_audio_mus_set_speed(OMG_Audio* this, OMG_Music* mus, float speed);
 #endif
